backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0